Refactor the dataexport stuff out of the main API class; it's borderline useless
